shell-ini-config: add ini_config_is_set() function
Due to the way that a variable is returned, it is not possible to
distinguish between a value that was set but is empty, or a value that
was never set. This prevents some possible logic on the configuration
file, including the ability to allow defaulting options. Some
configuration values might be legitimately set to null to (for example)
disable an option. This patch adds ini_config_is_set() function which
will return a 0 exit status if a match was found, and a non-zero exit
status if a match was not found. This allows determining the difference
between a set empty value versus a value which was not set at all, which
was not possible with ini_config_get() alone.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>